Generate unlock payload for pu_telnetEnabled
Netgear has changed how the unlock packet is generated.
The password is now sha256 hashed before being appended to the
plaintext and secret_key.
This commit implements the new scheme but breaks compatability
with older routers.
One important detail is that the secret_key is longer than allowable
than the blowfish standard.
This implementation does not validate key length so it works for
our purposes.
I believe it is the implementation you will find in Netgear GPL sources.